Fix bind/connect/getnameinfo/sendto issue because of bad addrlen.
In 4aa361d, the SOCK_ADDR_SIZE macro was removed and in bind(),
connect(), getnameinfo() and sendto() calls, it was replaced with
sizeof(struct sockaddr_storage). Unfortunately some systems (OSX) refuse
to work with an inexact (too large) addrlen.

Thanks to @sjthomason for reporting, bisecting the cause and reviewing
the fix. Closes #246.